# frontend/src/lib/api/tenant.openapi.json
{
  "components": {
    "schemas": {
      "AlertRuleIn": {
        "properties": {
          "alert_type": {
            "enum": [
              "below_reorder_point",
              "low_days_of_inventory",
              "batch_near_expiry",
              "expired_stock",
              "negative_stock",
              "count_variance",
              "food_cost_above_target",
              "approval_requested"
            ],
            "title": "Alert Type",
            "type": "string"
          },
          "channel": {
            "default": "in_app",
            "enum": [
              "in_app",
              "email"
            ],
            "title": "Channel",
            "type": "string"
          },
          "recipient_role_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient Role Id"
          },
          "recipient_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient User Id"
          }
        },
        "required": [
          "alert_type"
        ],
        "title": "AlertRuleIn",
        "type": "object"
      },
      "AlertRuleOut": {
        "properties": {
          "alert_type": {
            "enum": [
              "below_reorder_point",
              "low_days_of_inventory",
              "batch_near_expiry",
              "expired_stock",
              "negative_stock",
              "count_variance",
              "food_cost_above_target",
              "approval_requested"
            ],
            "title": "Alert Type",
            "type": "string"
          },
          "channel": {
            "default": "in_app",
            "enum": [
              "in_app",
              "email"
            ],
            "title": "Channel",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "recipient_role_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient Role Id"
          },
          "recipient_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient User Id"
          }
        },
        "required": [
          "alert_type",
          "id"
        ],
        "title": "AlertRuleOut",
        "type": "object"
      },
...
      "ApprovalRuleIn": {
        "properties": {
          "approver_role_id": {
            "format": "uuid",
            "title": "Approver Role Id",
            "type": "string"
          },
          "document_type": {
            "enum": [
              "purchase_order",
              "transfer",
              "adjustment",
              "count",
              "void",
              "refund",
              "discount",
              "journal"
            ],
            "title": "Document Type",
            "type": "string"
          },
          "min_amount": {
            "maximum": 1000000000000000.0,
            "minimum": 0.0,
            "title": "Min Amount",
            "type": "integer"
          },
          "outlet_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outlet Id"
          }
        },
        "required": [
          "document_type",
          "min_amount",
          "approver_role_id"
        ],
        "title": "ApprovalRuleIn",
        "type": "object"
      },
      "ApprovalRuleOut": {
        "properties": {
          "approver_role_id": {
            "format": "uuid",
            "title": "Approver Role Id",
            "type": "string"
          },
          "document_type": {
            "enum": [
              "purchase_order",
              "transfer",
              "adjustment",
              "count",
              "void",
              "refund",
              "discount",
              "journal"
            ],
            "title": "Document Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "min_amount": {
            "maximum": 1000000000000000.0,
            "minimum": 0.0,
            "title": "Min Amount",
            "type": "integer"
          },
          "outlet_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outlet Id"
          }
        },
        "required": [
          "document_type",
          "min_amount",
          "approver_role_id",
          "id"
        ],
        "title": "ApprovalRuleOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OutletOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "timezone",
          "is_active"
        ],
        "title": "OutletOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Restaurant POS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/alert-rules": {
      "get": {
        "operationId": "list_alert_rules_api_v1_alert_rules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AlertRuleOut"
                  },
                  "title": "Response List Alert Rules Api V1 Alert Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Alert Rules",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "create_alert_rule_api_v1_alert_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRuleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Alert Rule",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/alert-rules/{rule_id}": {
      "delete": {
        "operationId": "delete_alert_rule_api_v1_alert_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Alert Rule",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/approval-rules": {
      "get": {
        "operationId": "list_approval_rules_api_v1_approval_rules_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ApprovalRuleOut"
                  },
                  "title": "Response List Approval Rules Api V1 Approval Rules Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Approval Rules",
        "tags": [
          "settings"
        ]
      },
      "post": {
        "operationId": "create_approval_rule_api_v1_approval_rules_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalRuleIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApprovalRuleOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Approval Rule",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/approval-rules/{rule_id}": {
      "delete": {
        "operationId": "delete_approval_rule_api_v1_approval_rules__rule_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "rule_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Rule Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Approval Rule",
        "tags": [
          "settings"
        ]
      }
    },
...
    "/api/v1/outlets/{outlet_id}": {
      "get": {
        "description": "Another tenant's outlet is invisible (RLS); one outside the caller's scope is 404 too.",
        "operationId": "get_outlet_api_v1_outlets__outlet_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "outlet_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Outlet Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutletOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Outlet",
        "tags": [
          "tenant"
        ]
      }
    },
...
    "/api/v1/settings/{key}": {
      "put": {
        "operationId": "put_setting_api_v1_settings__key__put",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Put Setting Api V1 Settings  Key  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Setting",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}